sensu-go-check: Break out SensuGo as a generic class for future modules
